remove handler.py
