+++ .gitignore
@@ -1,3 +1,6 @@
+.vscode/
+.idea/
+
 /roverctl
 *.exe
 .DS_Store
